+++ src/DecoderError.ts
@@ -8,39 +8,39 @@ import * as FS from './FreeSemigroup'
  * @category model
  * @since 2.2.7
  */
-export interface Leaf {
+export interface Leaf<E> {
   readonly _tag: 'Leaf'
   readonly input: unknown
-  readonly error: string
+  readonly error: E
 }
 
 /**
  * @category model
  * @since 2.2.7
  */
-export interface Required {
+export interface Required<E> {
   readonly _tag: 'Required'
   readonly key: string
-  readonly errors: FS.FreeSemigroup<DecodeError>
+  readonly errors: FS.FreeSemigroup<DecodeError<E>>
 }
 
 /**
  * @category model
  * @since 2.2.7
  */
-export type DecodeError = Leaf | Required
+export type DecodeError<E> = Leaf<E> | Required<E>
 
 /**
  * @category constructors
  * @since 2.2.7
  */
-export const leaf = (input: unknown, error: string): DecodeError => ({ _tag: 'Leaf', input, error })
+export const leaf = <E>(input: unknown, error: E): DecodeError<E> => ({ _tag: 'Leaf', input, error })
 
 /**
  * @category constructors
  * @since 2.2.7
  */
-export const required = (key: string, errors: FS.FreeSemigroup<DecodeError>): DecodeError => ({
+export const required = <E>(key: string, errors: FS.FreeSemigroup<DecodeError<E>>): DecodeError<E> => ({
   _tag: 'Required',
   key,
   errors
@@ -50,11 +50,11 @@ export const required = (key: string, errors: FS.FreeSemigroup<DecodeError>): De
  * @category destructors
  * @since 2.2.7
  */
-export const fold = <R>(patterns: {
-  Leaf: (input: unknown, error: string) => R
-  Required: (k: string, errors: FS.FreeSemigroup<DecodeError>) => R
-}): ((e: DecodeError) => R) => {
-  const f = (e: DecodeError): R => {
+export const fold = <E, R>(patterns: {
+  Leaf: (input: unknown, error: E) => R
+  Required: (k: string, errors: FS.FreeSemigroup<DecodeError<E>>) => R
+}): ((e: DecodeError<E>) => R) => {
+  const f = (e: DecodeError<E>): R => {
     switch (e._tag) {
       case 'Leaf':
         return patterns.Leaf(e.input, e.error)
@@ -69,6 +69,6 @@ export const fold = <R>(patterns: {
  * @category instances
  * @since 2.2.7
  */
-export function getSemigroup(): Semigroup<FS.FreeSemigroup<DecodeError>> {
+export function getSemigroup<E = never>(): Semigroup<FS.FreeSemigroup<DecodeError<E>>> {
   return FS.getSemigroup()
 }